backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
anthropic==0.86.0
pymupdf==1.27.2.2
faiss-cpu==1.13.2
rank-bm25==0.2.2
sentence-transformers==5.3.0
numpy==2.4.2
python-dotenv==1.2.2
